logic: update filtered db names and schema dump
Post schema dump we retry to install the aiven_extras on the target db,
thus ensuring the extension is present even on new databases. Adding the
extension to the template* databases is not an option since there is no
discernable way to access that according to the pg_hba conf file we are
generating.

Add the _aiven database to the filtered db list along with the template*
databases. Migrating data to it would basically ensure the customer
loses access to that data on the target db, since there's way for a
normal user to access it

Database list will always retrieve all dbs from pg_catalog, thus fixing
any inconsistencies that may appear when migrating a cluster with more
DBs on the source.

Update default application name for better visibility on the upstream server