honour connect_timeout and PGCONNECT_TIMEOUT
The wait_select callback previously installed to enable Ctrl+C during
long queries was breaking configurable connection timeout :
https://github.com/psycopg/psycopg2/issues/944

This is replaced with a more visible async connection and a manual
call to a custom wait_select with support for timeout.

The timeout mimics default libpq behavior and reads the connect_timeout
connection parameter with a fallback on PGCONNECT_TIMEOUT environment
variable (and a default of 0: no timeout).

A secondary benefit is to allow importing PGMigrate inside another
project without PGMigrate altering the global set_wait_callback.